Update index.js
Hello Welly! First of all, thanks for your hard work.
Hello Welly! First of all, thanks for your hard work.
As to the problem, we have to return back status 404 to the client.
We should be setting the status on staticContext so that when rendering on the server we can access the status on the context object (routerContext.status) we passed.
https://reacttraining.com/react-router/web/guides/server-rendering/404-401-or-any-other-status